# storage/api-docs/api-docs.json
{"openapi":"3.0.0","info":{"title":"Task Resource OpenApi","contact":{"name":"\u5b66\u9662\u541b","url":"http:\/\/xueyuanjun.com","email":"support@todo.test"},"version":"3.0"},"servers":[{"url":"http:\/\/todo.test\/dingoapi\/tasks"}],"paths":{"\/":{"get":{"tags":["Tasks"],"summary":"Get list of tasks","description":"Returns list of tasks","operationId":"getTaskList","parameters":[{"name":"Accept","in":"header","description":"Accept header to specify api version","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","description":"The page num of the list","required":false,"schema":{"type":"integer"}},{"name":"limit","in":"query","description":"The item num per page","required":false,"schema":{"type":"integer"}}],"responses":{"200":{"description":"The result of tasks"}},"security":[{"passport":[]}]},"post":{"tags":["Tasks"],"summary":"Add New Task","description":"create new task","operationId":"newTaskItem","parameters":[{"name":"Accept","in":"header","description":"Accept header to specify api version","required":false,"schema":{"type":"string"}}],"requestBody":{"request":"text","description":"The text of the task","required":true},"responses":{"200":{"description":"The task item created","content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/task-transformer"}}}}},"security":[{"passport":[]}]}},"\/{id}":{"get":{"tags":["Tasks"],"summary":"Get Task","description":"Get specify task by id","operationId":"getTaskItem","parameters":[{"name":"Accept","in":"header","description":"Accept header to specify api version","required":false,"schema":{"type":"string"}},{"name":"id","in":"path","description":"The id of the task","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"The task item","content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/task-transformer"}}}},"404":{"description":"404 not found"}},"security":[{"passport":[]}]},"put":{"tags":["Tasks"],"summary":"Update Task","description":"update existed task by id","operationId":"updateTaskItem","parameters":[{"name":"Accept","in":"header","description":"Accept header to specify api version","required":false,"schema":{"type":"string"}},{"name":"id","in":"path","description":"The id of the task","required":true,"schema":{"type":"integer"}}],"requestBody":{"request":"task_in_body","description":"The task to update","required":true,"content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/task-model"}}}},"responses":{"200":{"description":"The task item updated","content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/task-transformer"}}}},"404":{"description":"404 not found"}},"security":[{"passport":[]}]},"delete":{"tags":["Tasks"],"summary":"Delete Task","description":"delete existed task by id","operationId":"deleteTaskItem","parameters":[{"name":"Accept","in":"header","description":"Accept header to specify api version","required":false,"schema":{"type":"string"}},{"name":"id","in":"path","description":"The id of the task","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"The task is deleted successful"},"404":{"description":"404 not found"}},"security":[{"passport":[]}]}}},"components":{"schemas":{"task-model":{"title":"Task model","properties":{"id":{"description":"The id of the task","type":"integer","format":"int64","example":1},"text":{"description":"The text of the task","type":"string","format":"string","example":"Test Task"},"is_completed":{"description":"If the task is completed or not","type":"string","format":"boolean","example":"1"}},"type":"object"},"task-transformer":{"title":"Task Transformer","properties":{"id":{"description":"The id of the task","type":"integer","format":"int64","example":1},"text":{"description":"The text of the task","type":"string","format":"string","example":"Test Task"},"completed":{"description":"If the task is completed or not","type":"string","format":"string","example":"yes"},"link":{"description":"The URL of the task detail page","type":"string","format":"string","example":"http:\/\/todo.test\/dingoapi\/task\/1"}},"type":"object"}},"securitySchemes":{"passport":{"type":"oauth2","description":"Laravel passport oauth2 security.","in":"header","scheme":"http","flows":{"password":{"authorizationUrl":"http:\/\/todo.test\/oauth\/authorize","tokenUrl":"http:\/\/todo.test\/oauth\/token","refreshUrl":"http:\/\/todo.test\/token\/refresh","scopes":[]}}}}}}